Changes to make cvr better

+++ cvr/main.go
@@ -2,9 +2,11 @@ package main
 
 import (
 	"flag"
+	"fmt"
 	"io/ioutil"
 	"log"
 	"os"
+	"strings"
 
 	"github.com/MediaMath/cove"
 	"github.com/pkg/browser"
@@ -32,25 +34,30 @@ func main() {
 	packs, pathErr := cove.Packages(paths...)
 	logError(pathErr)
 
-	profiles, coverErr := cove.CoverageProfile(*short, reportPath, packs...)
-	logError(coverErr)
+	anyCoverage := false
+	for _, pack := range packs {
+		profile, coverErr := cove.CoverageProfile(*short, reportPath, pack)
+		logError(coverErr)
 
-	if len(profiles) < 1 {
-		log.Printf("No coverage for %s, %s", paths, profiles)
+		if profile != "" {
+			if !*keepProfile {
+				defer os.RemoveAll(profile)
+			}
+
+			report, reportErr := cove.CoverageReport(profile, reportPath)
+			logError(reportErr)
+
+			if report != "" {
+				anyCoverage = true
+				if openReport {
+					browser.OpenFile(report)
+				}
+			}
+		}
 	}
 
-	for _, profile := range profiles {
-		if !*keepProfile {
-			defer os.RemoveAll(profile)
-		}
-
-		report, reportErr := cove.CoverageReport(profile, reportPath)
-		logError(reportErr)
-
-		if report != "" && openReport {
-			browser.OpenFile(report)
-		}
-
+	if !anyCoverage {
+		fmt.Printf("No coverage for %s\n", strings.Join(paths, ","))
 	}
 }
